Code review from phpcs
