+++ applicationSymfony/Controller/IdmlController.php
@@ -6,6 +6,7 @@ use AdventistCommons\Idml\DomManipulation\Exception as IdmlException;
 use App\Product\Idml\Validator;
 use App\Product\Form\Type\ValidateIdmlType;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
+use Symfony\Component\HttpFoundation\File\File;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
@@ -23,6 +24,7 @@ class IdmlController extends AbstractController
         $idmlValidationForm = $this->createForm(ValidateIdmlType::class);
         $idmlValidationForm->handleRequest($request);
         if ($idmlValidationForm->isSubmitted() && $idmlValidationForm->isValid()) {
+            /** @var File $file */
             $file = $idmlValidationForm->getData()['idmlFile'];
             $newPathname = $file->getFilename().'.idml';
             $file->move($file->getPath(), $newPathname);
