+++ applicationSymfony/Project/Form/Type/AddMemberType.php
@@ -2,12 +2,10 @@
 
 namespace App\Project\Form\Type;
 
-use App\Entity\Project;
 use App\Entity\User;
 use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class AddMemberType extends AbstractType
 {
